// src/@fake-db/types.ts
export type RecordPresensiType = {
  presense: string
  date: string,
  in: string,
  out: string,
  total: string,
}